requirements.txt
Flask==2.3.3
pandas==2.1.1
selenium==4.15.0
webdriver-manager==4.0.1
Werkzeug==2.3.7
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
flask-login==0.6.3
